Update workspace example to compile and be usable (#50)
A user on Buf public slack tried out the workspace example, however
prior to this PR the example generates code but fails on imports when
examining the generated code.

This PR updates the workspace example with a working minimal Go service
that uses the generated code from the workspace. The imports work
correctly using `go mod tidy` and readers can run the example with `go
run`.

This should be less confusing to users, and allows them to continue
working out the example for their own needs if they wish to do that.