Add Bufstream and Iceberg quickstart (#79)
This PR adds a quickstart for using Bufstream with Iceberg via a
Spark/Jupyter notebook. It's a companion for a docs PR submitted to the
same reviewers.

Notes on its contents:

1. Go, .proto, buf.yaml, and buf.gen.yaml are based on/copied from the
bufstream-demo repo. The only changes of note are in `producer.go`,
which:
    1.  Limit the number of messages to produce to 100.
    2. Omit trying to publish semantically invalid messages.
    3. Provide simpler log output.
2. The `docker-compose.yml` file used to coordinate containers is
annotated with each service's purpose.
3. Following reviewed/approved convention from Protovalidate examples,
its Makefile notes that it only exists for CI purposes. It's not used in
the documentation and the user does not need to have `make` available.

Notes on what I think it's missing:

I want to expand the code to include a "finish" state where an advanced
user can type `make`, wait for a success message, and immediately query
Iceberg via a notebook. I need to work with Bufstream engineers to find
a way to programmatically update the topic configuration without
additional dependencies.

---------

Co-authored-by: Carol Gunby <rakuista@users.noreply.github.com>
Co-authored-by: Chris Roche <github@rodaine.com>