Verify that all examples have READMEs linking back to their documentation (#90)
This PR adds missing READMEs to the buf-examples repository and applies
consistent language/formatting across all "new in 2025" readmes.

Key points:

1. I checked every directory.
2. I did not rewrite existing quickstarts for things like
`cli/managed-mode` - that's a larger content task that I think is best
approached the next time those quickstarts are revised in docs. I did
make the following changes within these older READMEs:
  1. GitHub action was missing its link to docs, and had a broken link.
  2. Remote plugin execution was missing its link to docs.
  3. Local plugin execution was missing its link to docs.
  4. Limiting types's link to docs was incorrect.

Questions for @rakuista:

1. In the examples that were moved over from Buf tour, the README refers
to "this folder," and then subdirectories as "directories." Is this
intentional, or should this all be "directory?"