Rework examples for predefined rules (#134)
The  nested `additional-examples` directory has additional examples
with changes represented in a sparse directory tree which can, in turn,
have its own `additional-examples` directory. Running `make ci` will
perform validations on the current directory, and recursively copy
itself and the rest of the boilerplate into the tree of
additional-examples directories to construct and then run validations
for each override scenario.

---------

Co-authored-by: Joe Rinehart <jrinehart@buf.build>